+++ Cargo.toml
@@ -52,15 +52,14 @@ branch = "feature/mqtt-interface"
 [dependencies.mcp23017]
 git = "https://github.com/mrd0ll4r/mcp23017.git"
 
-[dependencies.smoltcp]
-version = "0.6"
-features = ["ethernet", "proto-ipv4", "socket-tcp", "proto-ipv6"]
-default-features = false
+[dependencies.smoltcp-nal]
+git = "https://github.com/vertigo-designs/smoltcp-nal.git"
+branch = "main"
 
 [dependencies.stm32h7xx-hal]
 features = ["stm32h743v", "rt", "unproven", "ethernet", "quadspi"]
-git = "https://github.com/stm32-rs/stm32h7xx-hal"
-branch = "dma"
+git = "https://github.com/quartiq/stm32h7xx-hal"
+branch = "rs/smoltcp-update"
 
 [features]
 semihosting = ["panic-semihosting", "cortex-m-log/semihosting"]
